added the find path segment stuff of mla*. now need to connect that to lns to get the first iteration running

+++ src/instance.cpp
@@ -184,7 +184,7 @@ Instance::printAgents() const
         pair<int, int> start_loc = getCoordinate(start_locations[i]);
         PLOGI << "Agent " << i << " : S = (" << start_loc.first << ", " << start_loc.second
               << ") ;\nGoals : \n";
-        for (int j = 0; j < task_assignments[i].size(); j++) {
+        for (int j = 0; j < (int)task_assignments[i].size(); j++) {
             pair<int, int> goal_loc = getCoordinate(task_assignments[i][j]);
             PLOGI << "\t" << j << " : (" << goal_loc.first << " , " << goal_loc.second << ")\n";
         }
